[FIX] helpdesk: always send email to portal users
Since 7b338e0, Odoo does not send emails if you change the stage of your own
tickets (or create them), the rationale being that you don't need to be informed
of your own actions. But portal users are not familiar with Odoo so we want to
always send them an email if they create a ticket.

steps to reproduce:
- create a portal user
- create an helpdesk team with an alias
- set a template on the initial stage
- send an email from the portal user email address to that alias

before this commit:
- the ticket is created but no email is send from the mail template

after this commit:
- an email from the mail template is sent if a portal user creates a ticket
using the email alias

opw-3555716

closes odoo/enterprise#52091

X-original-commit: dd0c8e910f1fe44e9291624f8dc7d316524cfa9b
Signed-off-by: Nicolas Danhier (nda) <nda@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>