[FIX] account_accountant: bank reco: no linebreak in partner buttons
Translations may introduce whitespace into the text of some buttons.
The text of the buttons may then contain a linebreak.
This i.e. concerns the buttons in the "Manual Operations" tab that are visible
when a non-liquidity line is selected and has a partner set:
"Client:", "Payable:" below the partner.

Reproduce
  1. Select French language
  2. Go to the bank reconciliation widget and select an unmatched statement line.
  3. On the right side: Select any line that is not the liquidity line
  4. Ensure that the "Manual Operations" tab is selected
  5. Ensure that a partner is set and the 2 buttons below the partner are visible.
  6. The 2 buttons contain linebreaks.

closes odoo/enterprise#53668

X-original-commit: 06db316e554e5610bceb87760741386f33960607
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>