[FIX] account_bank_statement_import_csv: upload CSV file that has empty date
When the user imports CSV file of a bank statement and the file has
an empty date in one of the rows,
a traceback will appear.

Steps to reproduce the error:
- Go to Accounting > Dashboard > Bank > Import Statement
- Select a CSV file that has an empty date in one of the rows > Upload

Error: A traceback appears:
"TypeError: "<" not supported between instances of
 NoneType and datetime.date"

https://github.com/odoo/enterprise/blob/312b4e5df327b25e1678755b24bd27803720cd3d/account_bank_statement_import_csv/wizard/account_bank_statement_import_csv.py#L104-L105
Here, When one of the dates is empty, line[index_date] will be None
So, when it tries to sort dates at "sorted(dates)",
It will lead to above traceback.

sentry-4687473224

closes odoo/enterprise#58958

X-original-commit: a816367f5ece4f053c380e4dc38128018cd3da05
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>