[REF] spreadsheet_edition: rename revision_id field
The field `revision_id` is not a many2one field. It's a char field containing
a uuid string.

This field name is very confusing because every field ending with `_id` in
the codebase is a many2one field.
We also have some variables or parameters named `revision_id` which are meant
to be the real id (foreign key) of a revision.

The next step is to rename the client side "revisionId" and "nextRevisionId"
properties to match the matching model field names (resp. parent_revision_id
and revision_id). But this requires a bigger upgrade script.

Task: 3611266
Part-of: odoo/enterprise#51489