[FIX] industry_fsm: correct default project
Versions:
---------
- saas-16.2
- saas-16.3

Steps to reproduce:
-------------------
- create a new FSM project;
- create a task in this project;
- create a subtask for this task in the "Sub-tasks" tab.

Issue:
------
The project of the subtask is the "Field Service" project.

Cause:
------
If no project is found in the context, the first project found
in the default database is used by default.

Solution:
---------
Add the project of the parent task (found via the default context).
Consequently, we need to bypass the logic that looks for
a default project when we have a parent.

Note:
From version saas-16.4, no project is added.

opw-3602334

closes odoo/enterprise#52612

X-original-commit: 87313a56030bb47d04f38831ff1f1ad04b58ab20
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>