[IMP] web_studio: make UI of model configurator consistent
Before this commit, the style of the model configurator differs when being displayed in a Dialog,
or when being displayed directly from the app creator (which doesn't use the dialog component, and
don't have the appropriate styles applied by default). This commit changes the template to put the
title in the same layout as in the dialog, using <header> and <h4> elements, and reorganize some
scss around the ModelConfigurator component.

task-3572251

closes odoo/enterprise#49809

Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>