[FIX] *: default coa import error
*: l10n_ae_hr_payroll_account

Description of the issue/behavior this commit fixes:

When l10n_ae_hr_payroll_account is installed, trying to export the default CoA
and import it just after results in an error. This is due to the default
payroll account being set on a "Payable" type account while it should be an
"Expense" type one.

Steps to reproduce:

1 - Install l10n_ae_hr_payroll_account
2 - Use the AE Company
3 - Go to CoA, select all lines and export as importable data
4 - Go to import CoA, upload file, select the file of the previous step
5 - Test / Import the file
6 - An error says the import can't be performed

Desired behavior after this commit is merged:

This commit modifies the default payroll account to the "Basic Salary" one.
After this, importing the default CoA does not cause an error anymore as no
default account is being set to Payable anymore.

opw-3470011

closes odoo/enterprise#52997

X-original-commit: 5a4a106f129b14a0a6698fcbef3dfcfce450abf8
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>