[FIX] account_accountant: Duplicate reconcile button on cash journals
Description of the issue/behavior this commit fixes:

On the accounting dashboard, when at least one items is to be reconciled on a
cash journal, the box of that journal shows two reconcile buttons.

This behavior comes from opw-3384351 which was targetting 16.2 and was fw to
master while it should have stopped at 16.3. From 16.4, the reconcile button
is managed for cash journals.

Desired behavior after this commit is merged:

This commit reverses the problematic fix as from 16.4.

With this commit, only one button shows up at any time.

task-3624210

closes odoo/enterprise#52598

X-original-commit: f8c4acab1bf84ec4d80e4608175f9dd6b0bb12af
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>