[FIX] account_followup: partner follow-up level computation
To reproduce:
1- Create two bills, the first one should be +30 days overdue and the other between 15 and 30.
2- Execute manually the 15 days / level 1 follow-up. The partner goes into the 30 day level.
3- Change the Next Reminder date to the day before today.
4- Execute the second reminder manually..env['account_followup.followup.line'].search The partner status will go to 40 days.
5- Pay the first invoice.

Now when you go back to the follow-up report, the first invoice will have been removed (normal) but the status remains in "40 days" (not normal). This happens because the executed follow-up level is set for all the move lines, even the ones that are not due enough for the level.

Fix: the follow-up level set on the move lines is now the one having a delay equal to or less than the line's delay.

task-3608880

closes odoo/enterprise#53421

X-original-commit: 4569f0d9e20972dbed9007a6cd6b445e1aabf3df
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>