[FIX] l10n_si_reports: Fix wrong formulae in Balance Sheet and P&L
At the moment, Unaffected Earnings is a sub-line of line
VII. Carried forward net loss, which doesn't make any sense if the
unaffected earnings are profits.

In addition, separating profits and losses is inconsistent with the
official balance sheet of the Slovenian Accounting Standards (see task
for link.)

This commit therefore re-does the Retained Earnings / Current Year
Earnings section, to just keep 4 lines:
VI. Transferred net profit or loss
        VI.1 Retained Earnings / Carried forward net loss
        VI.2 Unaffected Earnings
VII. Net profit or loss for the financial year

In addition, in the Profit and Loss:
- Lines 3 (Decrease in the value of stocks of products and work in
  progress) and 10 (Operating Losses) should always be a positive value
  (consistent with 16 (Total Loss) and 20 (Net Loss for the period)
- Line 15 should be always be positive;
- Lines 19 and 20 were missing some terms.
This ensures that the Balance Sheet is balanced, which can be tested
using the Balance Sheet diagnosis tool in #36838.

closes odoo/enterprise#52461

Taskid: 3624482
X-original-commit: 519581f77db72e4a6135a6f40f88b56a682ccda1
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>