[FIX] web_enterprise: fix autofocus on the home menu
This commit fixes the behavior of the focus on the home menu when using
a touch oriented device. Before this commit, only the OS was considered,
and touch screen PCs always displayed the virtual keyboard when accessing
the home menu, since an invisible input was focused.

Steps to reproduce:
- on a Windows laptop or tablet with a touch screen, leave a view to go
back to the home menu
- the virtual keyboard appears
- you must touch out of the keyboard to use Odoo

Now, this input is no longer focused, and the user is not annoyed by the
keyboard popping up on the screen.

The hasTouch helper is used, since it can differentiate whether the laptop
is currently using the touch screen or the physical inputs.

task-3627697

closes odoo/enterprise#54116

X-original-commit: 26275d6583d39d1466025e9d4420b1fbcf39b8f5
Related: odoo/odoo#149024
Signed-off-by: Romain Estievenart (res) <res@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>