[FIX] whatsapp: save guests last seen message
Since 17.0, the new message separator is displayed for guests. The
set_last_seen_message rpc is also triggered in order to save this
information on the server side. However, the _channel_seen method
does not handle guests as it should.

As a consequence, the new message separator is always displayed and
the rpc call is always done when a guest loads the public page.

This PR enable guest support for this feature.

task-3603603

closes odoo/enterprise#52188

X-original-commit: b791be3c4b091e3be2496e0f86a86603b0ac680a
Related: odoo/odoo#145126
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>