[IMP] account_auto_transfer: Allow the archiving of transfers.
Following PR #52577, this commit adds the posibility to archive
automatic transfer which was not present so far.

Objective
---------
a) Allow transfers to be archived.
b) Update to optain the following behavior:
    1. If the transfer is linked to some POSTED account moves
        -> archive only
    2. If the transfer is linked to some DRAFT account moves
        -> archive OR
        -> delete the DRAFT moves, then delete the transfer
    3. Else, the delete and the archive actions should work right away.

Solution
---------
1. Add an active field to allow for the archiving of a transfer.
2. Add the search view to easily display the archived transfers and
update the form view to hide the buttons in archived transfers forms and
display the "Archived" ribbon when necessary.
3. Override the action_archive, so that archived transfers are
automatically disabled.
4. Addition of tests

task-3621529

closes odoo/enterprise#52115

Signed-off-by: John Laterre (jol) <jol@odoo.com>