[FIX] web_studio: select field through View All for Smart button
How to reproduce:
1) go to a form view
2) Open studio
3) Add Smart Button
4) Select a related field through the View All Dialog

Before this commit, the selection of field through View All when
adding a smart button on studio is not working.

After this commit, the update of the Many2XAutocompletion is reading the
display_name with an orm call. I reproduce the behaviour of the many2one_field.

Bugfix Task ID: 3618289

closes odoo/enterprise#52900

X-original-commit: 1329a424c343ae13dd0ff1aed4ebb04129c5b2b3
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>
Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>