[FIX] web_studio: select field through View All for Smart button
How to reproduce:
1) go to a form view
2) Open studio
3) Add Smart Button
4) Select a related field through the View All Dialog

Before this commit, the selection of field through View All when
adding a smart button on studio is not working.

After this commit, the update of the Many2XAutocompletion is reading the
display_name with an orm call. I reproduce the behaviour of the many2one_field.

Bugfix Task ID: 3618289

closes odoo/enterprise#52910

X-original-commit: 661249d3ed4cbcbeff4ed96f5cbc31d5307d55e2
Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>