[FIX] sale_subscription: hide plan_id when non-sub SO is confirmed
Currently, if you create a SO without subscription products and confirm it, the
plan_id field will still be editable. Therefore allowing to change a
confirmed SO to a subscription. This is an undesired behavior and should not be
possible. This commit fixes that by hiding the plan_id field in confirmed SO
without subscription products.

task-3600687

closes odoo/enterprise#52195

X-original-commit: 61a1a94f98dbcf4c203684066e40510937db1eaf
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>