[REF] *: user service is no longer a service
This commit adapts the enteprise codebase to a community change [1]
where the user service is removed and replaced by a simple object.

The `user` object can now be imported directly, without using
`useService` (in components) or declaring it as a dependency and
accessing it from the env (in services).

[1] https://github.com/odoo/odoo/pull/145362

Part of task 3621046

Part-of: odoo/enterprise#52372