[FIX] l10n_fr_reports: support french fiscal rounding
Each line of the french VAT report should be rounded to the unit. In
addition, the tax amounts should be recomputed using the rounded base
amounts!

The rounding of each line is enabled using the 'integer_rounding' option
introduced in
https://github.com/odoo/enterprise/commit/1916433a951fb98baf1a54e6b6e0f8a3f59920a4.
The recomputation of the tax lines base on the rounded base lines is
done using aggregation expression (so the `tax` tags are no longer used
in the report).

The rounding difference is accounted during the closing entry.

task-3607253

X-original-commit: 99851d4a6c05c69bb876104446b1046b74c3bf18
Part-of: odoo/enterprise#53341