[FIX] website_sale_subscription: display correct temporal unit website
Issue:
======
When having a product with duration variant and they have different
temporal units, it will only display the temporal unit of the first
selected variant in ecommerce product page.

Steps to reproduce the issue:
==============================
- Create a recurring product , add duration variant
- Add pricing for each variant using different periods
- Go to the product page in website and change the current variant , the
  temporal unit doesn't change.

Origin of the issue:
====================
It seems that the current code isn't reachable , and the function isn't
updated.
While debugging , this override of the function is never called but the
website_sale_stock one is called.

Solution:
=========
I copied the same format of code done in
website_sale_stock/static/src/js/variant_mixin.js and it solved the
issue.

opw-3625742

closes odoo/enterprise#55551

X-original-commit: e57a72a9ffa0912e014dc35512b9320d707c0016
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>