[FIX] quality: translate test type
quality.point.test_type names weren't being translated. Let's improve
the quality of the quality module for non-native English speakers!

Note that this fix is added into v15-saas-15.2, but could not be added
for v16-v17 stable versions due to adding "translate=True" to a Char
field requiring a -u of the module (since its translations are stored as
a different db column)

closes odoo/enterprise#52142

Signed-off-by: William Henrotin (whe) <whe@odoo.com>