[FIX] delivery_ups_rest: show UPS configuration tab
Steps to reproduce:
1. Install `delivery_ups_rest` module
2. Make a new shipping method
3. Choose "UPS" as provider

The UPS configuration tab that allows user to enter their credentials
and configuration will not show up. The problem is that the condition
for `invisible` attribute is defined in the old way. (vid https://github.com/odoo/enterprise/pull/35255)

This commit updates all of the modifiers in `delivery_ups_rest`.

opw-3618579

closes odoo/enterprise#52144

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>