[FIX] l10n_no_reports: Fix unbalanced Balance Sheet
The Balance Sheet was missing report lines specifically for Previous
Year Earnings and Current Year Earnings. This commit creates two new
report lines under Equity to represent them.

In addition, the transfer accounts 88 and 89 that are for setting the
Income / Expense accounts to zero at year end are now referenced in the
Balance Sheet rather than in the Profit and Loss. They should not affect
the Profit and Loss since they should normally only be used in the FY
closing entry.

Finally, account 835 Tax Cost was not referenced in the Profit and Loss,
so we added it.

This ensures that the Balance Sheet is balanced, according to the test
in https://github.com/odoo/enterprise/pull/36838.

closes odoo/enterprise#52858

Taskid: 3060790
X-original-commit: b1a0836fd6855663643eb32bf91e2d0d6cb3070e
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>