[FIX] l10n_at_reports: Using aggregation formula instead of domain formula
The aim of this commit is changing 2 lines where we are using
domain formula instead of aggregation formula.
The idea is that we don't want to have "redundant"
domains and we want to use the power of report engine by
using the right formula, in that case, the aggregation one.

closes odoo/enterprise#52620

Task-id: 3627640
X-original-commit: 0c895725d5f2afe4618504cdb85c6e4b90b72ecd
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>