[FIX] knowledge: do not mistakenly re-activate articles to archive
In the `_action_archive_articles` action, the `toggle_active` function
is called to deactivate all articles intended for archiving. Internally,
this function flips the active status of all records in the recordset.

If a user attempts to archive an already inactive article, the
`toggle_active` method inadvertently reactivates the article. Given the
SQL constraint stipulating that archived articles should be inactive,
users encounter a `CheckViolation` error when trying to archive an
inactive article.

To address this issue, we will invoke the `toggle_active` method on the
active records and disregard the inactive ones, ensuring that they are
not mistakenly reactivated.

task-3624569

closes odoo/enterprise#52686

X-original-commit: 165d216f76ea80b5d960b7bab8353760adac1831
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>