[IMP] core: allow to override model/field manual name rule
At Odoo, we want custom models and fields to start with `x_`.
However, other developers and companies might want to customize
that behavior and be able to change the rule.

This revision targets to factorize the rule `startswith('x_')`
in a dedicated method on `ir.model` and `ir.model.fields`
so they can be overridden to allow the customization
of custom model and fields name in custom modules.

closes odoo/enterprise#52279

X-original-commit: df6a53d4d3add9d486b10fee6c3a6379e7df3a9d
Related: odoo/odoo#145298
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>