[FIX] pos_pricer: neutralize
This commit adds the missing neutralisation necessary for the `pos_pricer`
module introduced in [1]

The purpose of the standard neutralisation framework is to allow us to
create database copies that will not interract with external systems in
ways that could impact the production database (or if it is not possible
to prevent the interractions, make sure that they are benign or wont
result in actual changes), or impact the customers of the operator of
the production database.

This is mainly useful to allow safe support investigation on database
duplicates.

[1] https://github.com/odoo/enterprise/pull/50368

closes odoo/enterprise#52217

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>