[FIX] account_accountant: Allow removal exchange diff in bank reco
In case the foreign currency is missing on the bank transaction, a partial could be suggested to the user
with an exchange difference. This commit allows the user to remove the exchange difference to spare the
generation of an exchange difference entry and the creation of a write-off line on the bank reco widget.

Suppose:
1200.0 comp_curr = 2400.0 foreign_curr in 2017 (rate 1:2)
1200.0 comp_curr = 3600.0 foreign_curr in 2016 (rate 1:3)
a statement line of 1200 comp_curr and an invoice of 3600.0 foreign_curr.

| Line       | foreign_curr | comp_curr |
| -----------| ------------ | --------- |
| liquidity  |              |    1200.0 |
| new_aml    |      -2400.0 |    -800.0 |
| exch diff  |              |    -400.0 |

...creating a partial reconcile on the invoice.
After removing the partial, you are forced to create a write-off:

| Line       | foreign_curr | comp_curr |
| -----------| ------------ | --------- |
| liquidity  |              |    1200.0 |
| new_aml    |      -3600.0 |   -1200.0 |
| exch diff  |              |    -600.0 |
| write-off  |              |     600.0 |

After this commit, you are able to remove the exchange difference:

| Line       | foreign_curr | comp_curr |
| -----------| ------------ | --------- |
| liquidity  |              |    1200.0 |
| new_aml    |      -3600.0 |   -1200.0 |

...allowing to fully reconcile the invoice without any additional exchange difference entry.

closes odoo/enterprise#52784

Task: 3628473
X-original-commit: 0531f7eef6902e48581ec0d05c01721f06e9a927
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>