[FIX] hr_appraisal_survey: fix 360 feedback default partner
Before this commit, answer's partner for 360 feedbacks were set
as the employee's user partner or the employee's private email.
This meant that all employees with no user would get the feedback request
through their private email instead of their work email.
However, it's not a wanted behaviour.

This commit changes the way the answer's partner is determined, taking
first the work partner, then the user partner if no work partner is found.

task-3627757 based on ticket-3612515

closes odoo/enterprise#55084

X-original-commit: 28361b936d6b3ef4df16a74b297362238e75db3a
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>