[FIX] social_twitter: re-activate users mentions
Purpose
=======
In 0cccfe3 , we automatically removed
the mentions if we are not mentioning the user we reply to, because we
had some issues with Twitter. Now that we have an enterprise
subscription, we want to re-activate them.

A config parameter has been added to allow disabling
mentions if necessary

Task-3615860

Part-of: odoo/enterprise#51752