[FIX] social_twitter: re-activate users mentions
Purpose
=======
In 0cccfe3 , we automatically removed
the mentions if we are not mentioning the user we reply to, because we
had some issues with Twitter. Now that we have an enterprise
subscription, we want to re-activate them.

A config parameter has been added to allow disabling
mentions if necessary

Task-3615860

X-original-commit: 85d2f6c97d0208c6c9840501af6aeec7773d424d
Part-of: odoo/enterprise#53607