[FIX] voip: make `how_to_call_on_mobile` editable on user.setting view
This commit makes `how_to_call_on_mobile` editable from the res.users
form.

Note: simply setting "readonly=0" resulted in a crash when no value is
provided for how_to_call_on_mobile, as it is a required field. Even
though it has a default value in res.users.settings, the value was not
properly forwarded to res.users. It turns out that related fields should
not be used with an inverse, causing all sort of odd behavior otherwise.
That's why this commit also replaces relateds with computes.

closes odoo/enterprise#52362

X-original-commit: ee3ae46aa7edc4f7719c136c26e5e297baf0b258
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>