[FIX] account_reports: unbalanced deferral due to rounding
When creating a deferral entry in the "Manual & Grouped" mode, we might
have a tiny rounding error on the final move, resulting in a user error
because of an unbalanced move. That way the user is blocked from
generating the deferral entry.

We now check before creating the deferral move whether the move is
balanced. If there is a tiny difference, we adjust the balance account
accordingly.

Part-of: odoo/enterprise#52489