[IMP] mrp_workorder: open backend MO in mrp_display
In the shop floor, only the MO had the button allowing them to open the
MO in the backend. The workorder now have it as well.

As we need to pass some context to the action to determine from where we started,
a new parameter "isMyWO" has been added to signify that the user in the action come
from the "My WO" page and need to go back on that page.

The "All MO" page is the default page, i.e. if the isMyWO is false and if there is no
workcenter.

closes odoo/enterprise#52258

Odoo: https://github.com/odoo/odoo/pull/145267
Task-id: 3624567
Related: odoo/odoo#145267
Signed-off-by: Steve Van Essche <svs@odoo.com>