[FIX] pos*: correct receipt printing for paid orders
Prior to this commit, attempting to print a receipt for a paid order in
the PoS restaurant resulted in an error. This issue was due to the
'getReceiptHeaderData' method in the PoS restaurant using
'this.get_order()' to retrieve the order, which is not applicable for
paid orders. Additionally, this problem exists in other modules that
override 'getReceiptHeaderData'.

To resolve this, the order is now passed as an argument to the
'getReceiptHeaderData' function.

opw-3616997

closes odoo/enterprise#52364

X-original-commit: fac8da312509c9b907f7a756f8ca9bddd7a1c7a2
Related: odoo/odoo#145499
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>