[FIX] pos*: correct receipt printing for paid orders
Prior to this commit, attempting to print a receipt for a paid order in
the PoS restaurant resulted in an error. This issue was due to the
'getReceiptHeaderData' method in the PoS restaurant using
'this.get_order()' to retrieve the order, which is not applicable for
paid orders. Additionally, this problem exists in other modules that
override 'getReceiptHeaderData'.

To resolve this, the order is now passed as an argument to the
'getReceiptHeaderData' function.

opw-3616997

closes odoo/enterprise#52249

Related: odoo/odoo#145252
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>