[FIX] account_accountant: compute empty deferred start date on post
Currently when you edit a posted move and remove the deferred start date
on a line, it won't be recomputed when posting again.
However we can not have a deferred end date without a deferred start
date. That's why we also compute the deferred start date when the state
of the move changes.

(Bug reported by TSB)

closes odoo/enterprise#52489

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>