[IMP] planning: shift templates improvements
This commit updates the method for calculating planned dates using shift
template values. It introduces the ability to modify the start and end
hours, as well as the duration in days, for which a shift template is
scheduled. The allocated hours is then determined using either the
user's or the company's calendar.

When assigning a shift template to a resource, the planned dates are
calculated according to the resource's calendar. The start and end hours
from the template are retained, along with the original start date. The
end date is generated based on the duration in days specified in the
template. The planned dates are arranged over a timeframe that ensures
the total working days match the specified days duration.

task-3624669

closes odoo/enterprise#52712

Related: odoo/upgrade#5438
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>