[FIX] industry_fsm: display planned date as a range in fsm task form
impacted modules:
- industry_fsm
- project_enterprise

Steps
=====
- install industry_fsm
- create a task in a fsm project

Issue
=====
The planned date field is not displayed as a range as it should.

Cause
=====
https://github.com/odoo/enterprise/pull/40866 introduces an onchange on
planned_date_begin and date_deadline that prevents to leave one of the
field empty while the other is set.

https://github.com/odoo/odoo/pull/143509 changes the behavior of the
daterange widget and display it as a single field if it is not
required.

Fix
===
A change to the daterange widget is introduced to force the display as a
range (see community pr).

task-3628069
related-https://github.com/odoo/odoo/pull/145167

closes odoo/enterprise#53589

X-original-commit: 9969bb7601e09b6d44f78c42726033cd643853de
Related: odoo/odoo#147928
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>