[IMP] l10n_br_avatax: make demoing tax calculation and EDI easier
This adds l10n_latam_identification_type_id to each demo partner so
there's less to configure. It also sets is_company, which was
forgotten before and makes more sense for these partners. It
additionally sets an address on BR Company Customer which allows easy
testing of in-state transactions (same state as the demo res.company).

We finally add some demo products that have been pre-configured to
test various tax and EDI calculation scenarios.

We also give a clear action to the user on how to resolve an error
received when trying to invoice service products.

task-3564171

closes odoo/enterprise#50780

Signed-off-by: Josse Colpaert <jco@odoo.com>