[FIX] knowledge: items aren't valid parents to move under
This commit fixes a bug where the article items were considered as valid
parents when using the **move to** button. These articles should not be
considered as such because you could easily lose them when moving them in
another hierarchy.

We changed the search domain inside the `get_valid_parent_options` function to
not include items.

task-3627789

closes odoo/enterprise#53884

X-original-commit: 5ec32cf8fe7527d8fc5569bb2a5bfa9ac580100e
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>