[FIX] l10n_us_1099: remove inconsistent demo data
Demo data of l10n_us_1099 used to create a bank statement that
was before the initial balance of `account` demo data which it
relies on. This caused inconsistencies in the balance shown in
the dashboard and the balance shown in the bank reco widget.
We remove this statement as anyway there are already enough data
provided with the accounting module.

closes odoo/enterprise#52463

Task-id: 3619563
X-original-commit: bced18c90c0c6ef37df0cfd212a2a540dd715ab5
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>