[FIX] web_studio: fix quick_create in kanban editor
This commit fixes the behavior of the 'Quick Create' attribute edition from the
view editor in Studio.

Before this commit, the value was not being read properly, and didn't showed the
correct value in the UI.

A test has been added, asserting the update of the checkbox value in the sidebar,
and the quick_create being edited via the edit_view call.

opw-3614654

closes odoo/enterprise#52541

X-original-commit: cb2fbae90ed8463f7eaa4130e7c05b0af6b8dc0c
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>