[FIX] web_studio: fix quick_create in kanban editor
This commit fixes the behavior of the 'Quick Create' attribute edition from the
view editor in Studio.

Before this commit, the value was not being read properly, and didn't showed the
correct value in the UI.

A test has been added, asserting the update of the checkbox value in the sidebar,
and the quick_create being edited via the edit_view call.

opw-3614654

closes odoo/enterprise#52629

X-original-commit: 6aa2151c5cd6de5c68991ea42495569060cb8ff1
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>