[IMP] account_reports: introduce integer rounding feature
In some countries, some reports (typically tax report or EC Sales List) need to be submitted to the state with rounded amounts (the exact rounding strategy possibly varying from one country to the other). So far, we did not support that.

In such cases, aggregations should be computed on the rounded values, so doing something in format_value was not enough. We solve this by introducing a new mechanism controlled from the options: integer rounding. This will be used in future commits to enforce such roundings, by calling the provided helper in a custom handler linked to the report, in _custom_options_initializer.

When the option is enabled, a new filter is displayed in the exra options, allowing to disable integer rounding. This way, users wishing to audit the details of the report don't loose precision. When printing the report, though, integer rounding is always enforced on reports setup to use it.

Note that using this option key on reports using a _dynamic_lines_generator will require case-by-case management of it within the custom code ; it's not supported automatically for them (since they don't use engines). This will change in master, but requires additional changes on the moment formatted values are computed (that are planned to land in master in the near future).

X-original-commit: 41d2b7a76ba92402c840b7b4b332ea16bbc3ec11
Part-of: odoo/enterprise#52337