[FIX] knowledge_frontend: title icon and text overlap (again)
Install website, knowledge and website_knowledge, open any knowledge
article and publish it to website visitor, then as a visitor access the
public link of that article. In case the article had a emoji on top,
that emoji and the title overlaps.

This is a partial revert of commit 7a8fde881229 which introduced those
changes in the layout to correct this same bug, also in 16.4, about the
title icon and the title text overlapping.

Reading the history, it seems that a generic frontend layout problem
arose in 16.4 in july, that 7a8fde881229 fixed the bug but in knowledge
only and that since then the generic frontend layout problem got
resolved, obsoleting 7a8fde881229 here in december.

References: 7a8fde881229 ([FIX] (website_)knowledge: fix various layout issues)
Task-3525618

closes odoo/enterprise#56439

X-original-commit: 6fd92635a9ba4ca1429aa2608a791f2d64d8acad
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>