[FIX] mrp_workorder: fix layout in mobile
This commit fixes an issue occuring in smaller screen sizes. The control panel
content is overlapping out of its parent, since no maximum width has been set to
the element. Since commit (1) redesigning the UI, the mw-100 and mw-lg-50 classes
have been removed, but setting a width must be maintained to avoid getting out
of the page.

This issue makes the page scroll to the right on mobile devices. The mw-100 class
has been reintroduced to maintain the right behavior, adding scrollbars when needed.

1) 580c78c90619b8265c42873b7fd16c42d5763703

closes odoo/enterprise#52359

X-original-commit: 9a2d990c3ba8354e8b7097bd5396b2a831a4e008
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>