[FIX] documents_spreadsheet: make contributors works with collaborative edition
Since 396fe57c08f3997a0a600190d12e2229089627bc, the spreadsheets shown
in the spreadsheet selector dialog (dialog to select a spreadsheet to
insert something into) are not correctly ordered anymore. This is due to
the fact that the contributor table is not correctly updated when a
modification is made in the spreadsheet, as with the collaborative edition
we use `spreadsheet.revision` to update the spreadsheet whereas before
we used the `write` method.

This commit fixes this issue by updating the contributor table when the
user join the spreadsheet.
We could also have updated the contributor table when the user do an
update (create a `spreadsheet.revision`), but this would have cause a
lot of useless updates.

Task-id 3601075

closes odoo/enterprise#54972

X-original-commit: 3529a9473917a464b1ad1c5d3814dae93950494e
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>