[IMP] appointment: adapt slot-creation mode colors
This commit improves the selected and non-clickable areas in the
calendar when the "Schedule availabilities" button is clicked.

The background colors of the `o_calendar_slot` used a green color
applied in JS which didn't render well.
This has been removed from the JS to apply the color in the SCSS
instead.

task-3630630
part of task-3575827

Part-of: odoo/enterprise#52344