[FIX] whatsapp: fix various composer / template usability issues
[FIX] hide 'phone' field in batch

This field does not add much value in batch mode: it displays list of
first 12 numbers that are about to be contacted. The only relevant
information is actually the number of invalid numbers, not the various
numbers themselves.

[FIX] template buttons

Improve error messages that are rather technical or hard to understand.
Also hide some buttons when the template is not approved: multi / action
buttons are used with validated templates.

Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: 7ff6e2b04110d1e3c046388927409ec46d39800e
Part-of: odoo/enterprise#52905